# openapi-appengine.yaml
...
swagger: "2.0"
info:
  description: "API for Quran"
  title: "Quran API"
  version: "1.0.0"
host: "utopian-button-227405.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
paths:
  "/v1/audio/{audioId}":
    get:
      tags:
        - "Audio"
      description: "Get Audio by Id"
      operationId: "getAudioById"
      parameters:
        - in: path
          name: audioId
          type: string
          required: true
          description: String ID of Audio
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
  "/v1/audio/ayah/{ayahId}":
    get:
      tags:
        - "Audio"
      description: "Get Audio by Ayah Id"
      operationId: "getAudioByAyahId"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
  "/v1/audio/edition/{editionId}":
    get:
      tags:
        - "Audio"
      description: "Get Audio by Edition Id"
      operationId: "getAudioByEdition"
      parameters:
        - in: path
          name: editionId
          type: string
          required: true
          description: String ID of Edition
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
  "/v1/audio/arabic/{ayahId}/{editionId}":
    get:
      tags:
        - "Audio"
      description: "Get Arabic Audio by Ayah Id and Edition Id"
      operationId: "getArabicAudio"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: path
          name: editionId
          type: string
          required: true
          description: String ID of Edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
  "/v1/audio/translation/{ayahId}/{editionId}":
    get:
      tags:
        - "Audio"
      description: "Get Audio Translation by Ayah Id and Edition Id"
      operationId: "getAudioTranslation"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: path
          name: editionId
          type: string
          required: true
          description: String ID of Edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
  "/v1/ayah/{ayahId}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Id"
      operationId: "getAyahById"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
...
  "/v1/edition/{editionId}":
    get:
      tags:
        - "Edition"
      description: "Get Edition by editionId"
      operationId: "getAyahByEdition"
      parameters:
        - in: path
          name: editionId
          type: string
          required: true
          description: ID of the edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Edition"
          schema:
            $ref: "#/definitions/edition"
...
definitions:
  audio:
    type: "object"
    properties:
      id:
        type: string
      ayahId:
        type: string
      ayahNumber:
        type: integer
      editionId:
        type: string
      type:
        type: string
        description: Type can be Audio or Translation
      link:
        type: string
        description: Link of the Audio file
  ayah:
    type: object
    properties:
      id:
        type: string
      surahId:
        type: string
      number:
        type: integer
      numberInSurah:
        type: integer
      juz:
        type: integer
      manzil:
        type: integer
      ruku:
        type: integer
      hizbQuarter:
        type: integer
      sajda:
        type: boolean
      arabic:
        type: string
  edition:
    type: object
    properties:
      id:
        type: string
      language:
        type: string
      translator:
        type: string
      type:
        type: string
      format:
        type: string
      direction:
        type: string
...
security: []